[ScheduleTreeTransform] Silence compiler warning. NFC.
Use size_t for position which is the return type type ArrayRef::size()
it is compared to.

llvm-svn: 362724

+++ polly/lib/Transform/ScheduleTreeTransform.cpp
@@ -439,7 +439,7 @@ struct ApplyASTBuildOptions
   BaseTy &getBase() { return *this; }
   const BaseTy &getBase() const { return *this; }
 
-  int Pos;
+  size_t Pos;
   llvm::ArrayRef<isl::union_set> ASTBuildOptions;
 
   ApplyASTBuildOptions(llvm::ArrayRef<isl::union_set> ASTBuildOptions)
